Add .siteMedia(blog:) source
