+++ WordPress/Classes/ViewRelated/Media/SiteMedia/SiteMediaPickerViewController.swift
@@ -7,7 +7,8 @@ protocol SiteMediaPickerViewControllerDelegate: AnyObject {
 
 /// The media picker for your site media.
 final class SiteMediaPickerViewController: UIViewController, SiteMediaCollectionViewControllerDelegate {
-    private let blog: Blog
+    let blog: Blog
+
     private let allowsMultipleSelection: Bool
     private let initialSelection: [Media]
